+++ html2text/config.py
@@ -57,7 +57,6 @@ USE_AUTOMATIC_LINKS = True
 # For checking space-only lines on line 771
 RE_SPACE = re.compile(r'\s\+')
 
-RE_UNESCAPE = re.compile(r"&(#?[xX]?(?:[0-9a-fA-F]+|\w{1,8}));")
 RE_ORDERED_LIST_MATCHER = re.compile(r'\d+\.\s')
 RE_UNORDERED_LIST_MATCHER = re.compile(r'[-\*\+]\s')
 RE_MD_CHARS_MATCHER = re.compile(r"([\\\[\]\(\)])")